可移植性: 标记指针函数 (tag_i/untag_i/to_int) 改用 intptr_t 实现 32/64 位兼容 — csrc/runtime.c: tag_i/untag_i/to_int 签名从 int32_t → intptr_t — csrc/runtime.c: ib/局部变量从 int32_t → intptr_t — csrc/runtime.c: printf %d → %lld (long long) 防止 64 位截断 — csrc/runtime.c: hash_key 中 untag_i 截断加显式 (uint32_t)(uintptr_t) 链 — csrc/runtime.c: strlen 转整数位数 buf[16]→buf[24], 加 (intptr_t) 宽适配 — llvmgen/runtime.c: 内嵌 VM 宏 _TAG_I/_UNTAG/_TO_INT 改用 intptr_t — llvmgen/runtime.c: 内嵌 VM ib/局部变量 + fprintf %d→%lld — gcc -std=c99 -Wall 0 警告, C VM 14/14 通过, 核心 292/292 通过

+++ llvmgen/runtime.c
@@ -1240,10 +1240,10 @@ void rt_throw(void *msg) {
 /* 前向声明 */
 void *rt_import(void *path);
 
-/* 标记指针 */
-#define _TAG_I(v)  ((void*)(intptr_t)(((int64_t)(v) << 1) | 1))
+/* 标记指针 ── 用 intptr_t 实现 32/64 位兼容 */
+#define _TAG_I(v)  ((void*)(intptr_t)(((intptr_t)(v) << 1) | 1))
 #define _IS_INT(p) (((intptr_t)(p) & 1) != 0)
-#define _UNTAG(p)  ((int32_t)((intptr_t)(p) >> 1))
+#define _UNTAG(p)  ((intptr_t)((intptr_t)(p) >> 1))
 #define _TO_INT(p) (_IS_INT(p) ? _UNTAG(p) : 0)
 
 /* 字节码 VM 结构 */
@@ -1373,7 +1373,7 @@ static int _bvm_run(_BVM *vm) {
         if (++steps > 50000000) { fprintf(stderr, "[rt_import] 超时\n"); return 1; }
         uint8_t op = _brd_u8(vm->code, &vm->pc);
         void *a, *b;
-        int32_t ib;
+        intptr_t ib;
 
         switch (op) {
         case 0x00: break; /* NOP */
@@ -1425,7 +1425,7 @@ static int _bvm_run(_BVM *vm) {
         case 0x14: b = _bpop(vm); a = _bpop(vm); _bpush(vm, _TAG_I(_TO_INT(a) < _TO_INT(b) ? 1 : -1)); break;
         case 0x15: b = _bpop(vm); a = _bpop(vm); _bpush(vm, _TAG_I(_TO_INT(a) >= _TO_INT(b) ? 1 : -1)); break;
         case 0x16: b = _bpop(vm); a = _bpop(vm); _bpush(vm, _TAG_I(_TO_INT(a) <= _TO_INT(b) ? 1 : -1)); break;
-        case 0x17: a = _bpop(vm); { int32_t v = _TO_INT(a); _bpush(vm, _TAG_I(v > 0 ? -1 : 1)); } break;
+        case 0x17: a = _bpop(vm); { intptr_t v = _TO_INT(a); _bpush(vm, _TAG_I(v > 0 ? -1 : 1)); } break;
 
         /* 字符串 */
         case 0x19: { /* CONCAT */
@@ -1476,7 +1476,7 @@ static int _bvm_run(_BVM *vm) {
 
         /* 输出 */
         case 0x0E: { a = _bpop(vm); /* PRINT - print to stderr for diagnostics */
-            if (_IS_INT(a)) fprintf(stderr, "%d\n", _UNTAG(a));
+            if (_IS_INT(a)) fprintf(stderr, "%lld\n", (long long)_UNTAG(a));
             else if (a) fprintf(stderr, "%s\n", ((rt_str_t*)a)->data);
             else fprintf(stderr, "null\n");
             break;
